[feat] #69 Change CardVIewModel
