+++ RollingPaper/RollingPaper/Controllers/Card/CardPencilKitViewController.swift
@@ -20,15 +20,15 @@ final class CardPencilKitViewController: UIViewController, PKCanvasViewDelegate,
     
     var backgroundImg = UIImage(named: "Rectangle")
     
-    private let viewModel: CardBackgroundViewModel
-    private let input: PassthroughSubject<CardBackgroundViewModel.Input, Never> = .init()
+    private let viewModel: CardViewModel
+    private let input: PassthroughSubject<CardViewModel.Input, Never> = .init()
     private var cancellables = Set<AnyCancellable>()
     
     private var isCanvasToolToggle: Bool = true
     private var isStickerToggle: Bool = true
     private var imageSticker: UIImage!
     
-    init(viewModel: CardBackgroundViewModel) {
+    init(viewModel: CardViewModel) {
         self.viewModel = viewModel
         super.init(nibName: nil, bundle: nil)
     }
@@ -112,6 +112,16 @@ final class CardPencilKitViewController: UIViewController, PKCanvasViewDelegate,
                 case .getRecentCardBackgroundImgFail:
                     DispatchQueue.main.async(execute: {
                         self.someImageView.image = UIImage(named: "Rectangle")
+                    })
+                case .getRecentCardResultImgSuccess(result: let result):
+                    DispatchQueue.main.async(execute: {
+                        //self.someImageView.image = UIImage(named: "Rectangle")
+                        print("test")
+                    })
+                case .getRecentCardResultImgFail:
+                    DispatchQueue.main.async(execute: {
+                       // self.someImageView.image = UIImage(named: "Rectangle")
+                        print("test")
                     })
                 }
             })
