+++ RollingPaper/RollingPaper/ViewModels/CardViewModel.swift
@@ -8,17 +8,21 @@
 import UIKit
 import Combine
 
-class CardBackgroundViewModel {
+class CardViewModel {
     var cardBackgroundImg = UIImage(named: "Rectangle")
+    var cardResultImg = UIImage(named: "heart.fill")
     
     enum Input {
         case viewDidLoad
         case setCardBackgroundImg(background: UIImage) //CardBackgroundViewController 으로부터 backgroundImgGet
+        case setCardResultImg(result: UIImage) //CardPencilKitViewController 으로부터 mergedImageSet
     }
     
     enum Output {
         case getRecentCardBackgroundImgSuccess(background: UIImage)
         case getRecentCardBackgroundImgFail
+        case getRecentCardResultImgSuccess(result: UIImage)
+        case getRecentCardResultImgFail
     }
     
     private let output: PassthroughSubject<Output, Never> = .init()
@@ -30,9 +34,13 @@ class CardBackgroundViewModel {
             switch event {
             case.viewDidLoad:
                 self.getRecentCardBackgroundImg()
+                self.getRecentCardResultImg()
             case.setCardBackgroundImg(let background):
                 self.setCardBackgroundImg(background: background)
                 self.getRecentCardBackgroundImg()
+            case.setCardResultImg(let result):
+                self.setCardResultImg(result: result)
+                self.getRecentCardResultImg()
             }
         })
         .store(in: &cancellables)
@@ -51,6 +59,21 @@ class CardBackgroundViewModel {
             output.send(.getRecentCardBackgroundImgSuccess(background: backImg!))
         } else {
             output.send(.getRecentCardBackgroundImgFail)
+        }
+    }
+    
+    private func setCardResultImg(result: UIImage) {
+        guard let png = result.pngData()
+        else { return }// png로 바꿔서 넣어 버린다.
+        UserDefaults.standard.set(png, forKey: "cardResultImg")
+    }
+    
+    private func getRecentCardResultImg() {
+        if let recentResultImg = UserDefaults.standard.data(forKey: "cardResultImg") {
+            let resultImg = UIImage(data: recentResultImg)
+            output.send(.getRecentCardResultImgSuccess(result: resultImg!))
+        } else {
+            output.send(.getRecentCardResultImgFail)
         }
     }
 }
